Move print vertex to generic package
