+++ keanu-project/src/main/java/io/improbable/keanu/vertices/Vertex.java
@@ -9,7 +9,7 @@ import io.improbable.keanu.network.NetworkLoader;
 import io.improbable.keanu.network.NetworkSaver;
 import io.improbable.keanu.tensor.Tensor;
 
-import io.improbable.keanu.vertices.utility.PrintVertex;
+import io.improbable.keanu.vertices.generic.nonprobabilistic.PrintVertex;
 import java.util.Arrays;
 import java.util.Collection;
 import java.util.Collections;
